refactor(logs): move LogSdkContext to @posthog/core
The type is only consumed through `PostHogLogs._getContext` (an SDK-internal
constructor argument). It was leaking into customer-facing
@posthog/types autocomplete despite never being a type customers should
write against.

Moves the definition to packages/core/src/logs/types.ts and re-exports
from @posthog/core. Browser SDK's types barrel re-imports from there.
